third_party/WebKit/LayoutTests/fast/events/touch/touch-event-source-device-event-sender.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<script src="../../../resources/js-test.js"></script>
<script type="text/javascript">
description('Test that touch events should have a sourceCapabilities and its firesTouchEvents should be true.');

function touchHandler() {
    shouldBeNonNull("event.sourceCapabilities");
    shouldBeTrue("event.sourceCapabilities.firesTouchEvents");
}

if (window.eventSender) {
    document.addEventListener("touchstart", touchHandler, false);
    document.addEventListener("touchmove", touchHandler, false);
    document.addEventListener("touchend", touchHandler, false);
    document.addEventListener("touchcancel", touchHandler, false);
  
    debug('Sending touchstart');
    eventSender.addTouchPoint(15, 15);
    eventSender.touchStart();

    eventSender.updateTouchPoint(0, 55, 55);
    debug('Sending touchmove');
    eventSender.touchMove();
    
    debug('Sending touchend');
    eventSender.releaseTouchPoint(0);
    eventSender.touchEnd();
    
    debug('Sending touchstart');
    eventSender.addTouchPoint(35, 35);
    eventSender.touchStart();
    
    debug('Sending touchcancel');
    eventSender.cancelTouchPoint(0);
    eventSender.touchCancel();
} else {
   debug('This test requires eventSender.');
}

</script>
