third_party/blink/web_tests/platform/generic/fast/events/touch/touch-event-source-device-event-sender-expected.txt

Test that touch events should have a sourceCapabilities and its firesTouchEvents should be true.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Sending touchstart
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is true
Sending touchmove
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is true
Sending touchend
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is true
Sending touchstart
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is true
Sending touchcancel
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is true
PASS successfullyParsed is true

TEST COMPLETE
